Second Big Push:: Update
This update features an upgraded spider for removing html tags, also added the url for each peace of content grouping for easy sorting. The file you want to inspect is the "life_extension_content.csv". See README for instructions for viewing the csv file in a data frame using panda.