Third Push:: Upgraded RegEx Filtering, Slice Adj
In this update, I changed the range of the slices to deal with the complex range of any of these articles. I also added some new regex expressions to handle dirty data. Last, check the life_extension_content.csv for accuracy, after spot checking I'm confident that we have the sub articles as well, the lines in the file are close to some 75K.