Protocol Sample Content Extraction
This is the csv file that output the data that was scraped. Due to the cluncky nature of internal linking and possible redirects, need to Q/A this file to see if indeed the accuracy of the pagination condition actually went through each rel nex/prev ux/ui element and gathered all the content.